// src/enums/monitor-plan-data-types.enum.ts
export enum MonitorPlanDataTypes {
  ANALYZER_RANGES = 'analyzer_range',
  ATTRIBUTES = 'monitor_location_attribute',
  COMPONENTS = 'component',
  DEFAULTS = 'monitor_default',
  DUCT_WAF = 'rect_duct_waf',
  FORMULAS = 'monitor_formula',
  LOADS = 'monitor_load',
  MATS_METHODS = 'mats_method_data',
  METHODS = 'monitor_method',
  QUALS_LEE = 'monitor_qualification_lee',
  QUALS_LME = 'monitor_qualification_lme',
  QUALS_PCT = 'monitor_qualification_pct',
  SPANS = 'monitor_span',
  SYSTEMS = 'monitor_system',
  SYSTEM_COMPONENTS = 'monitor_system_component',
  SYSTEM_FUEL_FLOWS = 'system_fuel_flow',
  UNIT_CAPACITY = 'unit_capacity',
  UNIT_CONTROL = 'unit_control',
  UNIT_FUEL = 'unit_fuel',
}

export default MonitorPlanDataTypes;